first commit with the changes related to the login / register and the refactoring related to the per user way

+++ bookstore/src/main/java/com/bookstore/bookstore/service/MyBookService.java
@@ -1,7 +1,7 @@
 package com.bookstore.bookstore.service;
 
 import java.util.List;
-
+import com.bookstore.bookstore.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,15 +13,23 @@ public class MyBookService {
 	@Autowired
 	MyBookRepository myBookRepo;
 
-	public void saveMyBooks(MyBookList b) {
-		myBookRepo.save(b);
+	public void saveMyBooks(MyBookList book) {
+		myBookRepo.save(book);
 	}
 
 	public List<MyBookList> getAllBook() {
 		return myBookRepo.findAll();
 	}
 
-	public void deleteById(int id) {
-		myBookRepo.deleteById(id);
+	public List<MyBookList> getAllMyBooks(User user) {
+		return myBookRepo.findByUser(user);
 	}
+
+	public void deleteById(int id, User user) {
+		MyBookList book = myBookRepo.findById(id).orElse(null);
+		if (book != null && book.getUser().getId() == user.getId()) {
+			myBookRepo.deleteById(id);
+		}
+	}
+
 }
